# Port `thrust::merge[_by_key]` to CUB