--- a/cub/cub/agent/agent_merge.cuh
+++ b/cub/cub/agent/agent_merge.cuh
@@ -181,7 +181,7 @@ struct agent_t
       item_type items_loc[items_per_thread];
       gmem_to_reg<threads_per_block, IsFullTile>(
         items_loc, items1_in + keys1_beg, items2_in + keys2_beg, num_keys1, num_keys2);
-      CTA_SYNC(); // TODO(bgruber): there is no sync when loading keys. bug or unnecessary?
+      CTA_SYNC(); // block_store_keys above uses shared memory, so make sure all threads are done before we write to it
       reg_to_shared<threads_per_block>(&storage.items_shared[0], items_loc);
       CTA_SYNC();