--- a/cub/cub/cub.cuh
+++ b/cub/cub/cub.cuh
@@ -64,6 +64,7 @@
 #include <cub/device/device_for.cuh>
 #include <cub/device/device_histogram.cuh>
 #include <cub/device/device_memcpy.cuh>
+#include <cub/device/device_merge.cuh>
 #include <cub/device/device_merge_sort.cuh>
 #include <cub/device/device_partition.cuh>
 #include <cub/device/device_radix_sort.cuh>