This PR ports the CUDA implementation of `thrust::merge[_by_key]` to CUB.

CUB already has an implementation of merge sort, also based on the merge path algorithm, so some infrastructure is taken from there, replacing the thrust implementation. However, whenever I was not sure what to pick, I preferred the thrust implementation and will attempt a unification in a separate PR after benchmarks are available. This especially concerns the agents.

The following new constructs are added:

* Public API entry points: `cub::DeviceMerge` with `MergeKeys` and `MergePairs`, including documentation.
* Internal dispatch layer: `cub::detail::DispatchMerge` with policy selection, fallback agent, vsmem implementation. Including a set of tuning policies taken over from the thrust implementation.
* Internal agent `cub::detail::AgentPartitionMergePath` implementing merge path partitioning.
* Internal agent `cub::detail::AgentMergeNoSort` implementing the device side merging. Very similar to merge sorts `AgentMerge` and I plan to unifying the implementations in a separate step.
* Extensive unit tests for the new CUB API.
* A few comments, drive-by fixes and refactorings.

A full tuning is probably necessary which will be conducted in a separate PR.

Compile-time of `thrust.test.merge`:
```
Before: 0m26.798s
After:  0m36.763s
```

<details>
<summary>Benchmark of `thrust.bench.merge` :</summary>

Running `nvbench_compare.py base.json branch.json`

```
## [0] NVIDIA H100 PCIe

|  T{ct}  |  Elements  |  Entropy  |  InputSizeRatio  |   Ref Time |   Ref Noise |   Cmp Time |   Cmp Noise |       Diff |   %Diff |  Status  |
|---------|------------|-----------|------------------|------------|-------------|------------|-------------|------------|---------|----------|
|   I8    |    2^16    |     1     |        25        |  21.842 us |       1.43% |  22.034 us |       5.27% |   0.192 us |   0.88% |   PASS   |
|   I8    |    2^20    |     1     |        25        |  27.602 us |       1.56% |  27.763 us |       2.02% |   0.161 us |   0.58% |   PASS   |
|   I8    |    2^24    |     1     |        25        |  66.583 us |       0.82% |  64.971 us |       0.88% |  -1.613 us |  -2.42% |   FAIL   |
|   I8    |    2^28    |     1     |        25        | 634.330 us |       0.31% | 613.761 us |       0.34% | -20.570 us |  -3.24% |   FAIL   |
|   I8    |    2^16    |   0.201   |        25        |  21.873 us |       1.78% |  22.052 us |       2.17% |   0.179 us |   0.82% |   PASS   |
|   I8    |    2^20    |   0.201   |        25        |  27.326 us |       3.29% |  27.322 us |       2.38% |  -0.004 us |  -0.02% |   PASS   |
|   I8    |    2^24    |   0.201   |        25        |  65.255 us |       0.64% |  63.658 us |       1.00% |  -1.597 us |  -2.45% |   FAIL   |
|   I8    |    2^28    |   0.201   |        25        | 630.302 us |       0.32% | 609.649 us |       0.36% | -20.653 us |  -3.28% |   FAIL   |
|   I8    |    2^16    |     1     |        50        |  22.411 us |       1.52% |  22.510 us |       3.64% |   0.099 us |   0.44% |   PASS   |
|   I8    |    2^20    |     1     |        50        |  28.515 us |       2.31% |  28.510 us |       1.29% |  -0.005 us |  -0.02% |   PASS   |
|   I8    |    2^24    |     1     |        50        |  67.800 us |       0.64% |  66.197 us |       0.66% |  -1.604 us |  -2.37% |   FAIL   |
|   I8    |    2^28    |     1     |        50        | 649.773 us |       0.46% | 629.247 us |       0.41% | -20.526 us |  -3.16% |   FAIL   |
|   I8    |    2^16    |   0.201   |        50        |  22.496 us |       2.60% |  22.547 us |       3.03% |   0.051 us |   0.23% |   PASS   |
|   I8    |    2^20    |   0.201   |        50        |  27.792 us |       1.56% |  27.946 us |       1.89% |   0.154 us |   0.56% |   PASS   |
|   I8    |    2^24    |   0.201   |        50        |  66.455 us |       0.60% |  64.921 us |       1.09% |  -1.534 us |  -2.31% |   FAIL   |
|   I8    |    2^28    |   0.201   |        50        | 637.270 us |       0.34% | 616.673 us |       0.38% | -20.597 us |  -3.23% |   FAIL   |
|   I8    |    2^16    |     1     |        75        |  22.178 us |       1.84% |  22.339 us |       2.51% |   0.161 us |   0.73% |   PASS   |
|   I8    |    2^20    |     1     |        75        |  28.112 us |       2.04% |  28.157 us |       1.54% |   0.045 us |   0.16% |   PASS   |
|   I8    |    2^24    |     1     |        75        |  66.745 us |       0.60% |  65.294 us |       1.18% |  -1.450 us |  -2.17% |   FAIL   |
|   I8    |    2^28    |     1     |        75        | 633.745 us |       0.31% | 612.842 us |       0.38% | -20.904 us |  -3.30% |   FAIL   |
|   I8    |    2^16    |   0.201   |        75        |  22.198 us |       2.97% |  22.337 us |       2.12% |   0.140 us |   0.63% |   PASS   |
|   I8    |    2^20    |   0.201   |        75        |  27.540 us |       2.01% |  27.694 us |       2.33% |   0.154 us |   0.56% |   PASS   |
|   I8    |    2^24    |   0.201   |        75        |  65.302 us |       0.94% |  64.032 us |       0.69% |  -1.270 us |  -1.94% |   FAIL   |
|   I8    |    2^28    |   0.201   |        75        | 626.013 us |       0.31% | 605.203 us |       0.35% | -20.810 us |  -3.32% |   FAIL   |
|   I16   |    2^16    |     1     |        25        |  23.348 us |       4.09% |  23.067 us |       1.57% |  -0.280 us |  -1.20% |   PASS   |
|   I16   |    2^20    |     1     |        25        |  30.352 us |       1.47% |  30.365 us |       1.18% |   0.013 us |   0.04% |   PASS   |
|   I16   |    2^24    |     1     |        25        |  82.447 us |       0.58% |  81.960 us |       0.47% |  -0.488 us |  -0.59% |   FAIL   |
|   I16   |    2^28    |     1     |        25        | 886.267 us |       0.25% | 879.144 us |       0.27% |  -7.123 us |  -0.80% |   FAIL   |
|   I16   |    2^16    |   0.201   |        25        |  23.129 us |       2.16% |  23.080 us |       1.97% |  -0.049 us |  -0.21% |   PASS   |
|   I16   |    2^20    |   0.201   |        25        |  30.261 us |       2.50% |  30.202 us |       1.33% |  -0.058 us |  -0.19% |   PASS   |
|   I16   |    2^24    |   0.201   |        25        |  75.287 us |       0.67% |  75.266 us |       0.66% |  -0.021 us |  -0.03% |   PASS   |
|   I16   |    2^28    |   0.201   |        25        | 748.692 us |       0.74% | 746.226 us |       0.78% |  -2.466 us |  -0.33% |   PASS   |
|   I16   |    2^16    |     1     |        50        |  23.782 us |       2.75% |  23.685 us |       2.17% |  -0.097 us |  -0.41% |   PASS   |
|   I16   |    2^20    |     1     |        50        |  30.969 us |       1.89% |  31.105 us |       1.34% |   0.137 us |   0.44% |   PASS   |
|   I16   |    2^24    |     1     |        50        |  84.724 us |       0.46% |  83.747 us |       0.93% |  -0.978 us |  -1.15% |   FAIL   |
|   I16   |    2^28    |     1     |        50        | 920.192 us |       0.19% | 901.230 us |       0.27% | -18.962 us |  -2.06% |   FAIL   |
|   I16   |    2^16    |   0.201   |        50        |  23.521 us |       3.51% |  23.697 us |       3.24% |   0.176 us |   0.75% |   PASS   |
|   I16   |    2^20    |   0.201   |        50        |  30.280 us |       1.76% |  30.764 us |       2.17% |   0.484 us |   1.60% |   PASS   |
|   I16   |    2^24    |   0.201   |        50        |  75.895 us |       0.59% |  76.160 us |       0.72% |   0.264 us |   0.35% |   PASS   |
|   I16   |    2^28    |   0.201   |        50        | 757.585 us |       0.77% | 755.773 us |       0.82% |  -1.812 us |  -0.24% |   PASS   |
|   I16   |    2^16    |     1     |        75        |  22.821 us |       2.00% |  23.086 us |       1.89% |   0.265 us |   1.16% |   PASS   |
|   I16   |    2^20    |     1     |        75        |  30.086 us |       1.32% |  30.490 us |       1.69% |   0.404 us |   1.34% |   FAIL   |
|   I16   |    2^24    |     1     |        75        |  82.047 us |       0.54% |  82.065 us |       0.57% |   0.018 us |   0.02% |   PASS   |
|   I16   |    2^28    |     1     |        75        | 886.565 us |       0.25% | 880.541 us |       0.28% |  -6.024 us |  -0.68% |   FAIL   |
|   I16   |    2^16    |   0.201   |        75        |  22.921 us |       1.86% |  23.136 us |       1.78% |   0.215 us |   0.94% |   PASS   |
|   I16   |    2^20    |   0.201   |        75        |  29.565 us |       1.29% |  30.108 us |       1.75% |   0.542 us |   1.83% |   FAIL   |
|   I16   |    2^24    |   0.201   |        75        |  75.021 us |       0.65% |  75.377 us |       0.79% |   0.355 us |   0.47% |   PASS   |
|   I16   |    2^28    |   0.201   |        75        | 744.530 us |       0.75% | 742.591 us |       0.80% |  -1.939 us |  -0.26% |   PASS   |
|   I32   |    2^16    |     1     |        25        |  23.432 us |       2.73% |  23.592 us |       2.23% |   0.159 us |   0.68% |   PASS   |
|   I32   |    2^20    |     1     |        25        |  31.594 us |       1.67% |  31.813 us |       1.21% |   0.219 us |   0.69% |   PASS   |
|   I32   |    2^24    |     1     |        25        | 103.510 us |       0.76% | 103.372 us |       0.64% |  -0.138 us |  -0.13% |   PASS   |
|   I32   |    2^28    |     1     |        25        |   1.263 ms |       0.18% |   1.262 ms |       0.14% |  -1.065 us |  -0.08% |   PASS   |
|   I32   |    2^16    |   0.201   |        25        |  24.005 us |       2.79% |  23.727 us |       3.65% |  -0.278 us |  -1.16% |   PASS   |
|   I32   |    2^20    |   0.201   |        25        |  31.509 us |       1.40% |  31.738 us |       1.17% |   0.229 us |   0.73% |   PASS   |
|   I32   |    2^24    |   0.201   |        25        | 103.135 us |       0.91% | 103.225 us |       0.77% |   0.090 us |   0.09% |   PASS   |
|   I32   |    2^28    |   0.201   |        25        |   1.244 ms |       0.12% |   1.245 ms |       0.12% |   1.002 us |   0.08% |   PASS   |
|   I32   |    2^16    |     1     |        50        |  24.019 us |       2.15% |  24.208 us |       2.70% |   0.189 us |   0.79% |   PASS   |
|   I32   |    2^20    |     1     |        50        |  32.874 us |       1.61% |  32.321 us |       1.25% |  -0.553 us |  -1.68% |   FAIL   |
|   I32   |    2^24    |     1     |        50        | 105.442 us |       0.78% | 104.331 us |       0.74% |  -1.111 us |  -1.05% |   FAIL   |
|   I32   |    2^28    |     1     |        50        |   1.280 ms |       0.19% |   1.279 ms |       0.16% |  -1.701 us |  -0.13% |   PASS   |
|   I32   |    2^16    |   0.201   |        50        |  25.211 us |       2.87% |  24.050 us |       2.10% |  -1.161 us |  -4.61% |   FAIL   |
|   I32   |    2^20    |   0.201   |        50        |  32.644 us |       2.14% |  32.125 us |       2.52% |  -0.519 us |  -1.59% |   PASS   |
|   I32   |    2^24    |   0.201   |        50        | 104.596 us |       0.73% | 104.011 us |       0.79% |  -0.586 us |  -0.56% |   PASS   |
|   I32   |    2^28    |   0.201   |        50        |   1.258 ms |       0.12% |   1.259 ms |       0.13% |   0.624 us |   0.05% |   PASS   |
|   I32   |    2^16    |     1     |        75        |  23.757 us |       2.02% |  23.200 us |       1.98% |  -0.557 us |  -2.34% |   FAIL   |
|   I32   |    2^20    |     1     |        75        |  32.017 us |       2.02% |  31.441 us |       1.17% |  -0.575 us |  -1.80% |   FAIL   |
|   I32   |    2^24    |     1     |        75        | 103.727 us |       0.80% | 102.959 us |       0.66% |  -0.768 us |  -0.74% |   FAIL   |
|   I32   |    2^28    |     1     |        75        |   1.264 ms |       0.20% |   1.261 ms |       0.14% |  -2.664 us |  -0.21% |   FAIL   |
|   I32   |    2^16    |   0.201   |        75        |  24.320 us |       2.78% |  23.597 us |       2.68% |  -0.723 us |  -2.97% |   FAIL   |
|   I32   |    2^20    |   0.201   |        75        |  32.081 us |       2.07% |  31.419 us |       1.12% |  -0.662 us |  -2.06% |   FAIL   |
|   I32   |    2^24    |   0.201   |        75        | 103.497 us |       0.77% | 102.911 us |       0.73% |  -0.585 us |  -0.57% |   PASS   |
|   I32   |    2^28    |   0.201   |        75        |   1.243 ms |       0.12% |   1.243 ms |       0.11% |   0.659 us |   0.05% |   PASS   |
|   I64   |    2^16    |     1     |        25        |  24.667 us |       2.80% |  24.259 us |       2.35% |  -0.408 us |  -1.66% |   PASS   |
|   I64   |    2^20    |     1     |        25        |  36.047 us |       1.21% |  35.505 us |       2.18% |  -0.543 us |  -1.51% |   FAIL   |
|   I64   |    2^24    |     1     |        25        | 179.325 us |       0.71% | 178.824 us |       0.63% |  -0.501 us |  -0.28% |   PASS   |
|   I64   |    2^28    |     1     |        25        |   2.508 ms |       0.10% |   2.507 ms |       0.07% |  -0.441 us |  -0.02% |   PASS   |
|   I64   |    2^16    |   0.201   |        25        |  25.525 us |       2.59% |  24.087 us |       2.99% |  -1.438 us |  -5.63% |   FAIL   |
|   I64   |    2^20    |   0.201   |        25        |  36.092 us |       1.31% |  35.179 us |       1.18% |  -0.913 us |  -2.53% |   FAIL   |
|   I64   |    2^24    |   0.201   |        25        | 178.690 us |       0.68% | 178.080 us |       0.68% |  -0.610 us |  -0.34% |   PASS   |
|   I64   |    2^28    |   0.201   |        25        |   2.474 ms |       0.06% |   2.474 ms |       0.06% |  -0.062 us |  -0.00% |   PASS   |
|   I64   |    2^16    |     1     |        50        |  25.163 us |       2.83% |  24.989 us |       2.70% |  -0.174 us |  -0.69% |   PASS   |
|   I64   |    2^20    |     1     |        50        |  37.091 us |       1.96% |  36.944 us |       1.84% |  -0.146 us |  -0.39% |   PASS   |
|   I64   |    2^24    |     1     |        50        | 181.273 us |       0.74% | 180.916 us |       0.71% |  -0.357 us |  -0.20% |   PASS   |
|   I64   |    2^28    |     1     |        50        |   2.548 ms |       0.12% |   2.546 ms |       0.08% |  -2.261 us |  -0.09% |   FAIL   |
|   I64   |    2^16    |   0.201   |        50        |  25.618 us |       3.17% |  25.167 us |       3.04% |  -0.450 us |  -1.76% |   PASS   |
|   I64   |    2^20    |   0.201   |        50        |  36.969 us |       1.10% |  36.682 us |       1.18% |  -0.287 us |  -0.78% |   PASS   |
|   I64   |    2^24    |   0.201   |        50        | 180.366 us |       0.73% | 180.155 us |       0.72% |  -0.211 us |  -0.12% |   PASS   |
|   I64   |    2^28    |   0.201   |        50        |   2.516 ms |       0.06% |   2.516 ms |       0.06% |   0.382 us |   0.02% |   PASS   |
|   I64   |    2^16    |     1     |        75        |  24.480 us |       2.20% |  24.336 us |       2.50% |  -0.144 us |  -0.59% |   PASS   |
|   I64   |    2^20    |     1     |        75        |  35.954 us |       1.26% |  35.794 us |       1.48% |  -0.160 us |  -0.45% |   PASS   |
|   I64   |    2^24    |     1     |        75        | 179.168 us |       0.70% | 178.804 us |       0.66% |  -0.365 us |  -0.20% |   PASS   |
|   I64   |    2^28    |     1     |        75        |   2.509 ms |       0.10% |   2.508 ms |       0.08% |  -1.246 us |  -0.05% |   PASS   |
|   I64   |    2^16    |   0.201   |        75        |  25.163 us |       3.20% |  24.693 us |       1.95% |  -0.469 us |  -1.87% |   PASS   |
|   I64   |    2^20    |   0.201   |        75        |  36.059 us |       1.94% |  35.744 us |       1.72% |  -0.314 us |  -0.87% |   PASS   |
|   I64   |    2^24    |   0.201   |        75        | 178.525 us |       0.65% | 178.218 us |       0.64% |  -0.307 us |  -0.17% |   PASS   |
|   I64   |    2^28    |   0.201   |        75        |   2.475 ms |       0.06% |   2.476 ms |       0.06% |   0.435 us |   0.02% |   PASS   |
|  I128   |    2^16    |     1     |        25        |  25.868 us |       1.83% |  25.704 us |       2.39% |  -0.164 us |  -0.63% |   PASS   |
|  I128   |    2^20    |     1     |        25        |  45.469 us |       1.38% |  45.148 us |       1.43% |  -0.321 us |  -0.71% |   PASS   |
|  I128   |    2^24    |     1     |        25        | 333.161 us |       0.32% | 333.008 us |       0.34% |  -0.154 us |  -0.05% |   PASS   |
|  I128   |    2^28    |     1     |        25        |   5.051 ms |       0.08% |   5.043 ms |       0.04% |  -8.013 us |  -0.16% |   FAIL   |
|  I128   |    2^16    |   0.201   |        25        |  26.753 us |       2.04% |  25.709 us |       1.69% |  -1.044 us |  -3.90% |   FAIL   |
|  I128   |    2^20    |   0.201   |        25        |  46.680 us |       1.72% |  46.337 us |       1.33% |  -0.342 us |  -0.73% |   PASS   |
|  I128   |    2^24    |   0.201   |        25        | 334.717 us |       2.16% | 334.350 us |       2.14% |  -0.368 us |  -0.11% |   PASS   |
|  I128   |    2^28    |   0.201   |        25        |   4.961 ms |       0.89% |   4.961 ms |       0.87% |  -0.137 us |  -0.00% |   PASS   |
|  I128   |    2^16    |     1     |        50        |  27.132 us |       2.11% |  26.928 us |       2.47% |  -0.204 us |  -0.75% |   PASS   |
|  I128   |    2^20    |     1     |        50        |  46.695 us |       1.25% |  46.189 us |       1.32% |  -0.506 us |  -1.08% |   PASS   |
|  I128   |    2^24    |     1     |        50        | 349.007 us |       1.83% | 348.394 us |       1.83% |  -0.614 us |  -0.18% |   PASS   |
|  I128   |    2^28    |     1     |        50        |   5.222 ms |       0.57% |   5.168 ms |       0.18% | -54.041 us |  -1.03% |   FAIL   |
|  I128   |    2^16    |   0.201   |        50        |  28.194 us |       3.60% |  27.417 us |       2.57% |  -0.778 us |  -2.76% |   FAIL   |
|  I128   |    2^20    |   0.201   |        50        |  47.612 us |       1.44% |  47.502 us |       1.90% |  -0.110 us |  -0.23% |   PASS   |
|  I128   |    2^24    |   0.201   |        50        | 342.038 us |       2.15% | 342.103 us |       2.17% |   0.065 us |   0.02% |   PASS   |
|  I128   |    2^28    |   0.201   |        50        |   5.072 ms |       0.74% |   5.072 ms |       0.88% |   0.732 us |   0.01% |   PASS   |
|  I128   |    2^16    |     1     |        75        |  26.335 us |       2.80% |  26.007 us |       1.66% |  -0.328 us |  -1.25% |   PASS   |
|  I128   |    2^20    |     1     |        75        |  45.377 us |       1.49% |  45.100 us |       1.43% |  -0.277 us |  -0.61% |   PASS   |
|  I128   |    2^24    |     1     |        75        | 339.315 us |       1.72% | 336.495 us |       1.85% |  -2.820 us |  -0.83% |   PASS   |
|  I128   |    2^28    |     1     |        75        |   5.067 ms |       0.10% |   5.052 ms |       0.10% | -14.962 us |  -0.30% |   FAIL   |
|  I128   |    2^16    |   0.201   |        75        |  27.758 us |       4.32% |  27.036 us |       1.78% |  -0.722 us |  -2.60% |   FAIL   |
|  I128   |    2^20    |   0.201   |        75        |  46.437 us |       1.21% |  46.439 us |       1.51% |   0.003 us |   0.01% |   PASS   |
|  I128   |    2^24    |   0.201   |        75        | 334.894 us |       2.10% | 334.996 us |       2.08% |   0.101 us |   0.03% |   PASS   |
|  I128   |    2^28    |   0.201   |        75        |   4.962 ms |       0.75% |   4.962 ms |       0.84% |  -0.037 us |  -0.00% |   PASS   |
|   F32   |    2^16    |     1     |        25        |  24.468 us |       2.21% |  23.690 us |       2.03% |  -0.778 us |  -3.18% |   FAIL   |
|   F32   |    2^20    |     1     |        25        |  31.909 us |       2.10% |  31.222 us |       1.67% |  -0.687 us |  -2.15% |   FAIL   |
|   F32   |    2^24    |     1     |        25        | 107.083 us |       1.02% | 104.880 us |       0.69% |  -2.203 us |  -2.06% |   FAIL   |
|   F32   |    2^28    |     1     |        25        |   1.294 ms |       1.44% |   1.272 ms |       2.79% | -21.351 us |  -1.65% |   FAIL   |
|   F32   |    2^16    |   0.201   |        25        |  28.905 us |       7.76% |  26.563 us |       6.99% |  -2.342 us |  -8.10% |   FAIL   |
…